books-steamgraph/requirements.txt
pandas
matplotlib
scipy
seaborn
PyYAML
numpy
